chore : change buildNFTAgentHal method location in file
move buildNFTAgentHal from handler_operation.go to handler_account.go
